+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bugfix
 - Fixed the axis title in bar chart - @pgee70 GH-267
+- PowerPoint2007 Writer : The presentation need repairs on Mac @jrking4 GH-266 GH-276
 
 ## 0.7.0 - 2016-09-12